# asg_form/bin/Release/net7.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ASG 赛事官网-后端API文档",
    "description": "这是由罗澜使用ASP.NET.Core开发的ASG赛事组后端系统，包括官网和后台管理系统。使用 sqlserver作为数据库，identity框架进行账号控制。",
    "version": "V 1.9.6"
  },
  "paths": {
...
    "/api/v1/form": {
      "put": {
        "tags": [
          "提交表单"
        ],
        "summary": "修改问卷",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "password",
            "description": "密码",
            "type": "string"
          },
          {
            "in": "query",
            "name": "formname",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "队伍信息",
            "schema": {
              "$ref": "#/definitions/form_get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      },
      "post": {
        "tags": [
          "提交表单"
        ],
        "summary": "提交表单",
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "captoken",
            "description": "谷歌人机验证验证码",
            "type": "string"
          },
          {
            "in": "body",
            "name": "body",
            "description": "表单信息",
            "schema": {
              "$ref": "#/definitions/form_get"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/api/v1/form/{team_name}": {
      "get": {
        "tags": [
          "提交表单"
        ],
        "summary": "搜索表单",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "team_name",
            "description": "表单名称",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/form"
            }
          }
        }
      }
    },
    "/api/v1/form/all": {
      "get": {
        "tags": [
          "提交表单"
        ],
        "summary": "获得所有表单信息",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "description": "页数",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "query",
            "name": "page_long",
            "description": "每页长度",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/team"
              }
            }
          }
        }
      }
    },
    "/api/v1/form/like": {
      "post": {
        "tags": [
          "提交表单"
        ],
        "summary": "点赞表单",
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "name",
            "description": "队伍名称",
            "type": "string"
          },
          {
            "in": "query",
            "name": "captoken",
            "description": "谷歌验证码token",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/like"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "form": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "piaoshu": {
          "format": "int32",
          "type": "integer"
        },
        "time": {
          "format": "date-time",
          "type": "string"
        },
        "team_name": {
          "type": "string"
        },
        "team_password": {
          "type": "string"
        },
        "team_tel": {
          "type": "string"
        },
        "logo_uri": {
          "type": "string"
        },
        "role": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/role"
          }
        }
      }
    },
    "form_get": {
      "type": "object",
      "properties": {
        "time": {
          "format": "date-time",
          "type": "string"
        },
        "team_name": {
          "type": "string"
        },
        "team_password": {
          "type": "string"
        },
        "team_tel": {
          "type": "string"
        },
        "logo_base64": {
          "type": "string"
        },
        "role_get": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/role_get"
          }
        }
      }
    },
...
    "like": {
      "type": "object",
      "properties": {
        "number": {
          "format": "int32",
          "type": "integer"
        }
      }
    },
...
    "role": {
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "form": {
          "$ref": "#/definitions/form"
        },
        "role_id": {
          "type": "string"
        },
        "role_name": {
          "type": "string"
        },
        "role_lin": {
          "type": "string"
        }
      }
    },
    "role_get": {
      "type": "object",
      "properties": {
        "role_id": {
          "type": "string"
        },
        "role_name": {
          "type": "string"
        },
        "role_lin": {
          "type": "string"
        }
      }
    },
    "roletwo": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "lin": {
          "type": "string"
        }
      }
    },
...
    "team": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "timer": {
          "format": "date-time",
          "type": "string"
        },
        "piaoshu": {
          "format": "int32",
          "type": "integer"
        },
        "logo_uri": {
          "type": "string"
        },
        "rolename": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/roletwo"
          }
        }
      }
    },
...
  }
}